[WIP]Merge 2.6 test to 2.6 release (#723)

+++ seatunnel-connectors-v2/connector-dolphindb-wt/src/test/java/org/apache/seatunnel/connectors/dolphindb/utils/DolphinDBSaveModeUtilTest.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.seatunnel.connectors.dolphindb.utils;
 
 import org.apache.seatunnel.api.table.catalog.Column;
@@ -18,11 +35,12 @@ public class DolphinDBSaveModeUtilTest {
     void fillingCreateSql() {
         List<Column> columns = new ArrayList<>();
 
-        columns.add(PhysicalColumn.of("id", BasicType.LONG_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("name", BasicType.STRING_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("age", BasicType.INT_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("gender", BasicType.BYTE_TYPE, null, true, null, ""));
-        columns.add(PhysicalColumn.of("create_time", BasicType.LONG_TYPE, null, true, null, ""));
+        columns.add(PhysicalColumn.of("id", BasicType.LONG_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("name", BasicType.STRING_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("age", BasicType.INT_TYPE, (Long) null, true, null, ""));
+        columns.add(PhysicalColumn.of("gender", BasicType.BYTE_TYPE, (Long) null, true, null, ""));
+        columns.add(
+                PhysicalColumn.of("create_time", BasicType.LONG_TYPE, (Long) null, true, null, ""));
 
         String result =
                 DolphinDBSaveModeUtil.fillingCreateSql(
